requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
cssselect==1.2.0
docker==6.1.3
idna==3.4
lxml==4.9.2
packaging==23.1
py-cone==1.1.2
PyMySQL==1.0.2
pyquery==2.0.0
redis==4.5.1
requests==2.28.2
urllib3==1.26.14
websocket-client==1.6.1
